worldwide-news-analytics/requirements.txt

elasticsearch==5.1.0
satori-sdk-python==1.0.2

